requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
langchain
pandas
numpy
python-dotenv
requests